Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (1 comment)

Karol
Should the parameter be in the general section or in each initiator section? I am thinking of each initiator section. We also got the input from someone else who is trying to evaluate the script running FIO to test RDMA at scale on over 500 systems. I think it's a good idea to provide the flexibility to configure this parameter but if there is not a configuration value provided we should default to the default fio.

Patch-set: 1
Reviewer: Gerrit User 1011314 <1011314@d5d70762-12d0-45a1-890d-524b12d3f735>