Update patch set 3
Patch Set 3:

This sounds good and I agree with you (maybe not on part how to solve the issue, but on the issue itself). But this is not a subject of this patch. Purpose of this patch is entirely different. I would suggest filling github issue for the issue you mention or creating Trello task.

Patch-set: 3
Reviewer: Gerrit User 1011262 <1011262@d5d70762-12d0-45a1-890d-524b12d3f735>