Update patch set 9
Patch Set 9:

> Patch Set 8: Code-Review-1
> 
> (3 comments)

Made requested changes and resubmitted, except for the comment about line 1008.
The reason the code is done the way it is here is because "opt" is optionally passed
and may be NULL.
The alternative, doing if (opt && ...) seems to me to be more code.
If this is the sole holdup, I will reluctantly comply.  :-)

Patch-set: 9
Reviewer: Gerrit User 1015860 <1015860@d5d70762-12d0-45a1-890d-524b12d3f735>