nvme example: hello_world
This change attempts to address the Trillo request to decode I/O errors in NVMe
hello_world example.

See https://trello.com/c/MzJJw7hM/2-decode-io-errors-in-nvme-helloworld-example

To test the failing paths, two compile time defines were added to force a
write or read error (bad LBA) respectively.

As the example does a read after write, if the write fails, the example fails.

In testing this, I noted that spdk_nvme_ns_get_id always returns an NSID of zero
if the namespace exists but is inactive.  This was temporarily addressed
in the example by passing the caller's NSID to the function (register_ns)
that displays inactive namespaces.

Signed-off-by: James Bergsten <jamesx.bergsten@intel.com>
Change-Id: Ib94b4a02495eb40966e3f49517a5bdf64485538a