[TEST]bdev: add uring in bdev Makefile
this patch is to discuss about liburing installing method.

We had "./configure --with-uring" , but actually, it wasn't compiled due to
there's a dependency of third party lib "liburing".
Test pool will not pass because of not installing "liburing".
If other one to edit bdev_uring, code will never check about this.

Change-Id: I8b98825f6795eb9f9e7b4947d1c7c3a44a6f0f64
Signed-off-by: yidong0635 <dongx.yi@intel.com>