Update patch set 5
Patch Set 5: Code-Review-1

On the second thought, the commit msg needs to be updated before we merge this patch.

I can still think of a scenario where it could be potentially useful, but the reason needs to be described in the commit msg. The current one is unfortunately wrong.

Patch-set: 5
Reviewer: Gerrit User 1011218 <1011218@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1