Update patch set 2
Patch Set 2:

Hi Ziye,

> I think that your code is ok. However, if we reset all the rc = SPDK_ISCSI_CONNECTION_FATAL. It will eliminate all the effective error info when we destroy this connection.

Now I could understand your idea, and it is good. the idea, that the original return code is returned to the caller and the caller moves the connection to exiting if the return code is negative, is nice.

Patch-set: 2
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>