Update patch set 3
Patch Set 3:

Change log is not ok. For us, we still have some kernels between 5.0 and 5.1. And if other customers use the old version or back porting the module without the optimized support, it will fail. I do not object setting the default value to 1, but if you do not export the config, customers need to change the spdk code to compile or change the kernel code in the initiator side, and it is not flexible.

Patch-set: 3
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>