dif: Factor out SGL append opeartion when the buffer is being splitted
The subsequent patches will refine spdk_dif_set_md_interleave_iovs
to change the data_len parameter to be the remaining length, and
to remove alignment constraint of the tail of the buffer. For the
head of the buffer any alignment has been supported. This patch is
a preparation for them and a little effort to make identical the
split case and the non-split case.

This patch doesn't change any behavior.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I0e9be7e66d313f3ec2bd8c55cce8bb18e4fff892