Update patch set 3
Patch Set 3:

Oh, this slipped past my radar. Issue #805 was closed two days after it was created and I haven't seen it before. Apparently we can hit this assertion with a standard AIO bdev, which is a bug somewhere in bdev_aio driver. I suggest to investigate the issue and revert this patch as soon as possible because it could lead to a series of other problems. Bdev modules must not return ENOMEM for the only pending I/O, because that I/O may never be re-sent.

Patch-set: 3
CC: Gerrit User 1011218 <1011218@d5d70762-12d0-45a1-890d-524b12d3f735>