Update patch set 4
Patch Set 4:

Tests failed due to Python formatting errors of some kind.

I think we don't need to support EOL kernels - customers aren't going to use those. However, now that the switch exists I'd like to keep it in. If anything, it's nice to be able to compare with it on vs off, and it may come in handy for debugging in the future.

Patch-set: 4
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>