event/app: don't call start_fn twice when json config is used
When JSON config was used, app layer was calling the
app start callback twice - once from internally-sent
"start_subsystem_init" RPC, and once from the app layer
itself.

To fix it, just forcibly set `--wait-for-rpc` option
whenever JSON config is used. It prevents the app from
being started from the app layer itself as it expects
it to be started from "start_subsystem_init" RPC. This
is exactly the case here.

I believe this was the orignal intent. The code to
override `--wait-for-rpc` was already there, but it
was setting it to a wrong value.

Fixes #816

Change-Id: Ib54d624f3167137216c910b2d947bbd1dc5023b1
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>