lib/reduce: clear bits for unused backing io unit array
When writing a chunk we set the unused io_unit_index[] values
to all F's but we weren't clearing the bits in the
allocated_backing_io_units bit array.

I'm not certain that this is a real problem or not but it's in
the area of where perf testing is running into its latest issue
as I'm seeing a bogus value in the io_unit_index[] array as
retrieved from getting the first clear bit in the backing array.

Change-Id: I5c4e5360c9c13817d2fa7d3c7d004b44537af7f4
Signed-off-by: paul luse <paul.e.luse@intel.com>