Update patch set 2
Patch Set 2:

After some thought I don't think this is causing issues but do believe it's good for clarity. We only ever partially write a chunk one time and at that time we allocate the backing io bits and set the array members accordingly. This is where we were not clearing the bits (however they weren't set). When the chunk is written to again it's not done in place and the code that 'retires' the old chunk does indeed clear the backing array bits.

Patch-set: 2
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Work-in-progress: false