nvme: check the support of Directive and enable it
If a controller supports Directives, then the controller shall:
1, Indicate support for Directives in the OACS field in the
Identify Controller data structure;
2, Support the Directive Receive command;
3, Support the Directive Send command;
4, Support the Identify Directive;

The Directive receive return parameters cmd returns a data
structure that contains a bit vector specifying the Directive
Types supported by the controller and a bit vector specifying the
Directive Types enabled for the namespace.
The Enable Directive operation is used to enable a specific
Directive for use within a namespace by all controllers that
are associated with the same Host Identifier.

Change-Id: Ib343bc34a9f00ab38b5afcfff3301622e1144d19
Signed-off-by: JinYu <jin.yu@intel.com>