Update patch set 16
Patch Set 16:

> Patch Set 16:
> 
> > Window Guest for vhost-blk will ignore the 'seg_max' field, so this is the issue I observed, by increasing the macro, it can fix #625.
> 
> Hi Changpeng,
> 
> What Darek says clearly and I say vaguely is that
> * did you try 128 parent_iovs and 32 child_iovs?
> * bdev layer has been implemented to work for 128 parent_iovs and 32 child_iovs.
> * If we see any issue for 128 parent_iovs and 32 child_iovs, we should fix it first.
> * Then as an optimization, we may be able to increase BDEV_IO_NUM_CHILD_IOV to 128.
> 
> Could you take the first three steps?

The main reason for this issue is that not all the iov buffer is 512 bytes aligned, for this case we need the backend's decision, if the backend is NVMe which has hardware SGL support, it's fine to accept the request, then if the backend is AIO, it will go through spdk_bdev_io_get_buf() again to copy all the original iovs into a new data buffer. I hope the explanation is clear to you now, :).

Patch-set: 16
Reviewer: Gerrit User 1011204 <1011204@d5d70762-12d0-45a1-890d-524b12d3f735>