Update patch set 17
Patch Set 17:

> You understanding is wrong, if we increase the value of BDEV_IO_NUM_CHILD_IOV, the total count of each request is 512 Bytes aligned, this is already being guaranteed. The error "Remaining X is not a multiple of block size Y" can only happen if we are running out of children iov, because IO boundary must be 512 aligned. Are you clear for now?

I think Changpeng is correct. When we run out child iovs, we will submit child IO to backend, and so each child IO transfer must be a multiple of block size.

But I think we can implement another solution.

Another solution may be that when we detect run-out of child iov, we step back or rewind to the offset to make the size multiple of block size.

I think we may be able to implement the feature to rewind to the offset the size of child IO is multiple of block size.

Patch-set: 17
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>