Update patch set 4
Patch Set 4: Code-Review-1

This is really complex, but I think we need to solve this in a bit broader way. When a namespace is added or removed from a subsystem, the subsystem is first paused. That means all new I/O is queued up by their poll groups and it waits for all outstanding I/O to complete before proceeding. Once the pause operation is done, the subsystem and the controllers can be freely updated by the subsystem thread because nothing else is happening on any other thread.

Except there's one complication around sending AERs. The poll groups are still running because they may be handling connections for other subsystems. So when the AER is sent from the subsystem thread (which calls spdk_nvmf_request_complete), if that AER is acknowledged very quickly the poll group that owns the admin queue (the ctrlr thread) may begin processing the acknowledgement at the same time that spdk_nvmf_request_complete is still executing on the subsystem thread. This is the bug.

I think the best fix is actually going to be an update to the way the subsystem is resumed so that it not only sends a message to each poll group, but also to each controller to notify that the resume has occurred. In that message, we can send the AERs.

Patch-set: 4
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1