Update patch set 10
Patch Set 10:

>> According to your idea, we should abandon all the test cases with bdev via using applications (iSCSI, NVMe-oF, vhost). So your reason with -2 cannot convince me.

This is exactly our direction. We only put in combination tests of (network protocol) + (bdev) when we don't have tests that run directly on the bdev layer that hit the problematic code path. And even then, all of those tests should be considered temporary.

So if what you are saying is that there isn't a specific scenario that iSCSI + crypto is hitting that we can't cover via bdevperf or fio with the bdev plugin, then the direction is to not add this test.

Patch-set: 10
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>