ut/cunit: redefine asserts to *_FATAL variants
This patch is so that we terminate execution of test case,
when failure occurs. Otherwise static analyzers will report
issues when continuing on clearly incorrect execution.

Only alternative would be to change a slew of required asserts
to fatal ones in code, probably volunurable to missing some.
This apporach is surefire way to get them all.

This is related to issue #822.

Change-Id: I372c92c7110705505b06120eda3f6639773282d3
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>