Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Actually I'd like to do similar stuff on other RPC calls as well. Changes related to: https://github.com/spdk/spdk/issues/319

yeah, that's why I made the comment and put the -1 to make sure everyone is on the same page. If we change the return val for every bdev RPC then everyone out there who is using them is going to get a really nasty surprise when they rebase if these are merged this way.  One recent example, the RAID RPC had a parm in it that didn't make sense and in this case I knew hardly anyone was using it but we (everyone) still decided to take the extra steps necessary to avoid surprising someone we might not know about.

I'd suggest we get a good solid comm (and maintainer) agreement on this apprach vs make all of these new APIs with slight name changes or something.

Patch-set: 6
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>