[RFC] doc/lvol.md: make the doc more user friendly
The logical volumes documentation is currently something
inbetween a user guide and a programmer guide. It's not
helpful as either of those, and today it's hard to understand
how lvols can be used - especially when it comes to more
complex features like snapshots and clones. Since there's
no other user guide for lvols, we'll now make this doc/lvol.md
to be directed strictly towards end users. This implies:

* remove all low-level bits, point to blobstore documentation
  for technical details
* describe lvols as bdevs, separate lvols could be described
  in an lvol programmer's guide, but we don't have one today
* give a brief overview of how logical volumes work
* provide RPC call examples and describe their output - show
  how lvol management looks like

Change-Id: Icea3bfb0adc7b57e243aa8e97ea459d45b12b1ea
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>