test/iscsi_tgt: remove drive size dependency
As we add more virtual machines for testing, we may want to decrease the
average size of emulated NVMe drives to conserve disk space. The current
ext4test in iscsi presupposes that the provided NVMe drive is at least
20 GiB. This patch removes this assumption but keeps the cap of 10 GiB
per split for the test.

Note: the reason for the 10 GiB max is so that large drives don't take
an incredibly long time to create the filesystem for very large drives.

Change-Id: I4035905de42d07bdcef795f39b0ccf4cb076c31f
Signed-off-by: Seth Howell <seth.howell@intel.com>