Update patch set 8
Patch Set 8: Code-Review-1

(1 comment)

These stats are going to be really useful, thanks!

It makes sense that you have to make the pending stat cumulative. I don't actually see a way around that other than changing the spdk_bdev_io_free api function to return an int or adding a new API call like spdk-bdev_io_free_verbose that returns the number of pending I/O it submitted when 
I'm not really opposed to adding a new API function if you wanted to make this a real time statistic, but you might want to poll some of the architects on the project about that.

I still see the cumulative statistic as valuable, but there needs to be some documentation about it. And if possible, you should change the name of the variable that gets printed out to the user to something like cum_pending_bdev_ios or something to tip the user off that it's a cumulative stat that needs to be analyzed relative to other iterations of the stats variable.

Patch-set: 8
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1